Restructure pyvista -> graphic
