+++ doc/source/api/visualization/pathlines.rst
@@ -3,4 +3,4 @@
 Pathlines
 =========   
 
-.. autopostdoc:: ansys.fluent.visualization.pyvista.pyvista_objects.Pathlines
+.. autopostdoc:: ansys.fluent.visualization.graphic.graphic_objects.Pathlines
