+++ doc/source/api/visualization/plots.rst
@@ -22,7 +22,7 @@ a monitor plot.
         
         #To plot data on local surface created in PyVista
         
-        from ansys.fluent.visualization.pyvista import  Graphics        
+        from ansys.fluent.visualization import  Graphics
         pyvista_surface_provider = Graphics(session).Surfaces       
         plots_session = Plots(session, pyvista_surface_provider)
         plot2 = plots_session.XYPlots["plot-2"]
