+++ src/ansys/fluent/visualization/graphic/__init__.py
@@ -1,0 +1,6 @@
+"""A package that provides interfacing Fluent with graphic renderer (PyVista)."""
+
+from ansys.fluent.visualization.graphic.graphic_objects import Graphics  # noqa: F401
+from ansys.fluent.visualization.graphic.graphic_windows_manager import (  # noqa: F401
+    graphic_windows_manager,
+)
