+++ src/ansys/fluent/visualization/graphic/graphic_windows_manager.py
@@ -1,0 +1,12 @@
+"""Module for graphic windows management."""
+
+from ansys.fluent.visualization.graphic.pyvista.graphic_defns import (
+    PyVistaWindowsManager,
+)
+
+
+class GraphicWindowsManager(PyVistaWindowsManager):
+    pass
+
+
+graphic_windows_manager = GraphicWindowsManager()
